examples/05-kafka-iot-integration-v2/requirements.txt
paho-mqtt
kafka-python
